Option to provide both redisClient and subscriber client
For users using alternative redis libs (e.g. ioredis + cluster),
this allows the caller to provide both the primary redisClient as well
as an additional client instance for the redis subscriber. This removes
the responsibility of this module to create any redis connections,
giving the caller flexibility to use whatever redis lib they want.